Import SSKeychain.
SSKeychain is a Cocoapods project which provides handy wrappers around
the iOS keychain (which has a Core Foundation-based API). This import
pulls it in but adjusts all its exported symbols' prefix from "SS" to
"LPSS" to avoid duplicate symbols with any projects that want to link
Calabash and use SSKeychain themselves.