Make: add xctests rule from running XCTest
The xct remains, but is now an alias for xctests.

FYI/BTW, 'make xctest' was my first choice, but this conflicts with
make's behavior:  if a directory with same name as the rule exists
the rule will exit "xctest is up to date".  The XCTests are in a
directory called 'XCTest'.